# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevFolio AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  subject?: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export interface GithubScanInput {
  /** @minLength 1 */
  username: string;
}

export interface Repository {
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  language: string | null;
  stars: number;
  forks: number;
  url: string;
  topics?: string[];
  hasReadme?: boolean;
  hasDeployment?: boolean;
  /** Quality score 0-100 */
  quality?: number;
}

export interface GithubScanResult {
  username: string;
  avatarUrl: string;
  /** @nullable */
  bio: string | null;
  totalRepos: number;
  totalStars: number;
  followers?: number;
  following?: number;
  topLanguages: string[];
  repositories: Repository[];
  contributionStreak?: number;
  /** Portfolio readiness score 0-100 */
  portfolioReadiness?: number;
}

export type AiSuggestionType = typeof AiSuggestionType[keyof typeof AiSuggestionType];


export const AiSuggestionType = {
  readme: 'readme',
  screenshot: 'screenshot',
  deployment: 'deployment',
  description: 'description',
  skills: 'skills',
  project: 'project',
} as const;

export type AiSuggestionPriority = typeof AiSuggestionPriority[keyof typeof AiSuggestionPriority];


export const AiSuggestionPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface AiSuggestion {
  type: AiSuggestionType;
  title: string;
  description: string;
  priority: AiSuggestionPriority;
  /** @nullable */
  project?: string | null;
}

export interface AiSuggestionsResult {
  username: string;
  suggestions: AiSuggestion[];
  strongestSkills: string[];
  /** Overall portfolio score 0-100 */
  overallScore: number;
}

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  company: string;
  avatar: string;
  content: string;
  rating: number;
}

export interface PlatformStats {
  portfoliosGenerated: number;
  githubProfilesScanned: number;
  developersSatisfied: number;
  avgGenerationTime: string;
}