# lib/api-zod/src/generated/types/aiSuggestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevFolio AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiSuggestionPriority } from './aiSuggestionPriority';
import type { AiSuggestionType } from './aiSuggestionType';

export interface AiSuggestion {
  type: AiSuggestionType;
  title: string;
  description: string;
  priority: AiSuggestionPriority;
  /** @nullable */
  project?: string | null;
}