# lib/api-zod/src/generated/types/aiSuggestionsResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevFolio AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiSuggestion } from './aiSuggestion';

export interface AiSuggestionsResult {
  username: string;
  suggestions: AiSuggestion[];
  strongestSkills: string[];
  /** Overall portfolio score 0-100 */
  overallScore: number;
}