# lib/api-zod/src/generated/types/githubScanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevFolio AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubScanInput {
  /** @minLength 1 */
  username: string;
}