# lib/api-zod/src/generated/types/githubScanResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevFolio AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Repository } from './repository';

export interface GithubScanResult {
  username: string;
  avatarUrl: string;
  /** @nullable */
  bio: string | null;
  totalRepos: number;
  totalStars: number;
  followers?: number;
  following?: number;
  topLanguages: string[];
  repositories: Repository[];
  contributionStreak?: number;
  /** Portfolio readiness score 0-100 */
  portfolioReadiness?: number;
}