# lib/api-zod/src/generated/types/repository.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevFolio AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Repository {
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  language: string | null;
  stars: number;
  forks: number;
  url: string;
  topics?: string[];
  hasReadme?: boolean;
  hasDeployment?: boolean;
  /** Quality score 0-100 */
  quality?: number;
}